src/pyproject.toml
[tool.poetry]
name = "django-react"
version = "0.1.0"
description = "A Simple Django backend with react frontend using typescript and vite"
authors = ["Arvind <pingarvindforquestions@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = ">=4.2,<4.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
whitenoise = "^6.6.0"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
djlint = "^1.34.1"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
